rearranged files and use statements
